Application and lambda added
Nodes Application and Lambda added as per #495.
Lambda has exactly one parameter. Application
has an arbitrary number of arguments applied
to an expression.

Nested applications are combined into one when
inner simplified. Then, there's a switch over
built-in functions (like `sin`, ..., `derivative`,
etc.), which either get simplified into a built-in
node, or are left as they are if there was not
enough arguments provided.

Lambda does not get simplified in any way, unless
something is applied to it. In that case, the given
argument substitutes the lambda's parameter. If
there is a conflict between the substituted value
and local parameter of another lambda, the inner
lambda gets alpha-converted to one of unused
variables.