Tan bug fixed (#570)
So basically, before that `tan` checked if the angle is divided
by `pi / 2`. If yes, it should be zero.

But it misses a condition: it also should not be integer. Basically,
whenever you get `pi / 2` as a modulo of dividing the angle by
`pi`, then `tan` turns into NaN. Whereas it just checked dividing
by `pi / 2` and did not care about case when `angle` is `pi` or
any integer by `pi`.

Issue #568