Infinite recursion in integral fix (#613)
Whenever you do InnerSimplified against an integral node, it would try to evaluate the integral. If it stays an integral, we do not evaluate again.

However, in certain cases it would not return an integral - but it would return something like a sum of integrals. In that case, we don't want to evaluate anymore either.